services/payments/requirements.txt
# reqs
Flask==2.3.2
psycopg2-binary==2.9.9
requests==2.32.3
gunicorn==21.2.0
